layout fix
The widest content of some boxes would spill outside the box on narrow (not-mobile) page widths. This fix will make each box twice as wide on those widths. Also added some "empty" divs that make sure the boxes are all the same height, so that the 4x1 row becomes a pretty 2x2 grid.